requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.6,<3.0
python-multipart>=0.0.9,<1.0
pillow>=10.3,<12.0
httpx>=0.27,<1.0
sqlalchemy>=2.0,<3.0
asyncpg>=0.29,<1.0
aiosqlite>=0.20,<1.0
chromadb>=0.5,<1.0
